+++ CSETWebApi/CSETWeb_Api/BusinessLogic/BusinessLogic.csproj
@@ -367,6 +367,7 @@
     <Compile Include="BusinessManagers\Diagram\analysis\NetworkZone.cs" />
     <Compile Include="BusinessManagers\Diagram\layers\LayerManager.cs" />
     <Compile Include="BusinessManagers\DocumentManager.cs" />
+    <Compile Include="BusinessManagers\EdmNistCsfMapping.cs" />
     <Compile Include="BusinessManagers\FrameworkManager.cs" />
     <Compile Include="BusinessManagers\FullAnswer.cs" />
     <Compile Include="BusinessManagers\InformationTabData\BaseQuestionInfoData.cs" />
