+++ CSETWebNg/src/app/reports/edm/edm-appendix-a-1/edm-appendix-a-1.component.html
@@ -27,7 +27,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[0].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[0].Acronym)">
           {{append_a_data[0].FunctionName}} ({{append_a_data[0].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -124,7 +124,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[0].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[0].Acronym)">
           {{append_a_data[0].FunctionName}} ({{append_a_data[0].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -237,7 +237,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[0].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[0].Acronym)">
           {{append_a_data[0].FunctionName}} ({{append_a_data[0].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -295,7 +295,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[1].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[1].Acronym)">
           {{append_a_data[1].FunctionName}} ({{append_a_data[1].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -392,7 +392,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[1].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[1].Acronym)">
           {{append_a_data[1].FunctionName}} ({{append_a_data[1].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -450,7 +450,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[1].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[1].Acronym)">
           {{append_a_data[1].FunctionName}} ({{append_a_data[1].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -536,7 +536,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[1].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[1].Acronym)">
           {{append_a_data[1].FunctionName}} ({{append_a_data[1].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -594,7 +594,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[2].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[2].Acronym)">
           {{append_a_data[2].FunctionName}} ({{append_a_data[2].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -692,7 +692,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[2].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[2].Acronym)">
           {{append_a_data[2].FunctionName}} ({{append_a_data[2].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -750,7 +750,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[3].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[3].Acronym)">
           {{append_a_data[3].FunctionName}} ({{append_a_data[3].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -876,7 +876,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[3].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[3].Acronym)">
           {{append_a_data[3].FunctionName}} ({{append_a_data[3].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -962,7 +962,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[4].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[4].Acronym)">
           {{append_a_data[4].FunctionName}} ({{append_a_data[4].Acronym}})
         </div>
         <div class="edm-aa-right-content">
