+++ CSETWebNg/src/app/reports/edm/edm-appendix-a-2/edm-appendix-a-2.component.html
@@ -22,7 +22,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[0].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[0].Acronym)">
           {{append_a_data[0].FunctionName}} ({{append_a_data[0].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -91,7 +91,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[0].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[0].Acronym)">
           {{append_a_data[0].FunctionName}} ({{append_a_data[0].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -179,7 +179,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[0].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[0].Acronym)">
           {{append_a_data[0].FunctionName}} ({{append_a_data[0].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -268,7 +268,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[0].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[0].Acronym)">
           {{append_a_data[0].FunctionName}} ({{append_a_data[0].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -355,7 +355,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[0].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[0].Acronym)">
           {{append_a_data[0].FunctionName}} ({{append_a_data[0].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -429,7 +429,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[1].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[1].Acronym)">
           {{append_a_data[1].FunctionName}} ({{append_a_data[1].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -499,7 +499,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[1].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[1].Acronym)">
           {{append_a_data[1].FunctionName}} ({{append_a_data[1].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -585,7 +585,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[1].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[1].Acronym)">
           {{append_a_data[1].FunctionName}} ({{append_a_data[1].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -671,7 +671,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[1].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[1].Acronym)">
           {{append_a_data[1].FunctionName}} ({{append_a_data[1].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -729,7 +729,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[2].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[2].Acronym)">
           {{append_a_data[2].FunctionName}} ({{append_a_data[2].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -827,7 +827,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[2].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[2].Acronym)">
           {{append_a_data[2].FunctionName}} ({{append_a_data[2].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -885,7 +885,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[3].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[3].Acronym)">
           {{append_a_data[3].FunctionName}} ({{append_a_data[3].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -1011,7 +1011,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[3].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[3].Acronym)">
           {{append_a_data[3].FunctionName}} ({{append_a_data[3].Acronym}})
         </div>
         <div class="edm-aa-right-content">
@@ -1097,7 +1097,7 @@
         <div class="edm-aa-header edm-body-col-4">EDM References</div>
       </div>
       <div class="edm-aa-body-split">
-        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFramgeworkColor(append_a_data[4].Acronym)">
+        <div class="edm-aa-left-id edm-body-col-1" [ngStyle]="getFrameworkColor(append_a_data[4].Acronym)">
           {{append_a_data[4].FunctionName}} ({{append_a_data[4].Acronym}})
         </div>
         <div class="edm-aa-right-content">
