+++ CSETWebNg/src/app/reports/edm/edm-appendix-a-2/edm-appendix-a-2.component.ts
@@ -71,7 +71,7 @@ export class EDMAppendixASectionTwo implements OnInit {
     return retVal
   }
   
-  getFramgeworkColor(input){
+  getFrameworkColor(input){
     let color = "rgb(0,0,0)"
     switch(input){
       case 'ID': {
